Fix two warnings/errors.  Inerestingly, only shiraz (RHEL5) gave a warning/error, tikal (RHEL6) did not.
- Fix a cast.
- Change a variable from long to double


git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@6635 640d5228-2204-0410-8d71-8f46fa6850b4

+++ instruments/pms2d/process2d/process2d.cpp
@@ -623,7 +623,7 @@ int process2d(Config & cfg, netCDF & ncfile, ProbeInfo & probe)
       * all of them.
       */
      {
-     long newbuffertime = TwoDtime(&buffer) + ((double)ntohs(buffer.msec) / 1000);
+     double newbuffertime = TwoDtime(&buffer) + ((double)ntohs(buffer.msec) / 1000);
      if (newbuffertime != buffertime)
      {
        buffertime = newbuffertime;
@@ -635,7 +635,7 @@ int process2d(Config & cfg, netCDF & ncfile, ProbeInfo & probe)
      if (buffcount == 0)
      {
        firstday = ntohs(buffer.day);
-       last_time1hz = buffertime;
+       last_time1hz = (long)buffertime;
        continue;
      }
